natlink_compat: fold _ui_dispatch into _ui_protocol
