+++ tests/test_ui_dispatch.py
@@ -1,4 +1,4 @@
-"""test_ui_dispatch.py - Unit tests for natlink_compat._ui_dispatch.
+"""test_ui_dispatch.py - Unit tests for natlink_compat._ui_protocol.
 
 Tests state snapshot building, phase setting, and UI provider notification.
 Does NOT require Dragon.
@@ -19,7 +19,7 @@ class TestBuildStateSnapshot(unittest.TestCase):
         self._state.reset()
 
     def test_disconnected_state(self):
-        from natlink_compat._ui_dispatch import build_state_snapshot
+        from natlink_compat._ui_protocol import build_state_snapshot
         from natlink_compat._ui_protocol import PHASE_IDLE
 
         state = build_state_snapshot()
@@ -27,19 +27,19 @@ class TestBuildStateSnapshot(unittest.TestCase):
         self.assertEqual(state.phase, PHASE_IDLE)
         self.assertEqual(state.mic_state, "")
         self.assertEqual(state.user_name, "")
-        self.assertEqual(state.loaders, ())
+        self.assertEqual(state.loader_states, ())
 
     def test_connected_state(self):
-        from natlink_compat._ui_dispatch import build_state_snapshot
+        from natlink_compat._ui_protocol import build_state_snapshot
         from natlink_compat._ui_protocol import PHASE_CONNECTED
 
         mock_backend = MagicMock()
         mock_backend.dragon_version = (15, 0, 0)
 
         self._state.backend = mock_backend
-        self._state._current_phase = PHASE_CONNECTED
-        self._state._last_mic_state = "on"
-        self._state._last_user_name = "TestUser"
+        self._state.phase = PHASE_CONNECTED
+        self._state.last_mic_state = "on"
+        self._state.last_user_name = "TestUser"
 
         state = build_state_snapshot()
         self.assertTrue(state.connected)
@@ -49,7 +49,7 @@ class TestBuildStateSnapshot(unittest.TestCase):
         self.assertEqual(state.dragon_version, (15, 0, 0))
 
     def test_snapshot_is_frozen(self):
-        from natlink_compat._ui_dispatch import build_state_snapshot
+        from natlink_compat._ui_protocol import build_state_snapshot
 
         state = build_state_snapshot()
         with self.assertRaises(AttributeError):
@@ -67,19 +67,19 @@ class TestSetPhase(unittest.TestCase):
         self._state.reset()
 
     def test_set_phase_updates_state(self):
-        from natlink_compat._ui_dispatch import set_phase
+        from natlink_compat._ui_protocol import set_phase
         from natlink_compat._ui_protocol import PHASE_CONNECTING
 
         set_phase(PHASE_CONNECTING)
-        self.assertEqual(self._state._current_phase, PHASE_CONNECTING)
+        self.assertEqual(self._state.phase, PHASE_CONNECTING)
 
     def test_set_phase_with_error(self):
-        from natlink_compat._ui_dispatch import set_phase
+        from natlink_compat._ui_protocol import set_phase
         from natlink_compat._ui_protocol import PHASE_ERROR
 
         set_phase(PHASE_ERROR, "connection failed")
-        self.assertEqual(self._state._current_phase, PHASE_ERROR)
-        self.assertEqual(self._state._error_message, "connection failed")
+        self.assertEqual(self._state.phase, PHASE_ERROR)
+        self.assertEqual(self._state.error_message, "connection failed")
 
 
 class TestNotifyUI(unittest.TestCase):
@@ -94,7 +94,7 @@ class TestNotifyUI(unittest.TestCase):
         self._state.reset()
 
     def test_notify_ui_calls_provider(self):
-        from natlink_compat._ui_dispatch import notify_ui
+        from natlink_compat._ui_protocol import notify_ui
 
         mock_provider = MagicMock()
         self._state.ui_provider = mock_provider
@@ -105,7 +105,7 @@ class TestNotifyUI(unittest.TestCase):
         self.assertFalse(state_arg.connected)
 
     def test_notify_ui_survives_provider_error(self):
-        from natlink_compat._ui_dispatch import notify_ui
+        from natlink_compat._ui_protocol import notify_ui
 
         bad_provider = MagicMock()
         bad_provider.on_state_changed.side_effect = RuntimeError("boom")
@@ -113,7 +113,7 @@ class TestNotifyUI(unittest.TestCase):
         notify_ui()  # Should not raise
 
     def test_no_providers_is_noop(self):
-        from natlink_compat._ui_dispatch import notify_ui
+        from natlink_compat._ui_protocol import notify_ui
         self._state.ui_provider = None
         notify_ui()  # Should not raise
 
@@ -130,7 +130,7 @@ class TestNotifyText(unittest.TestCase):
         self._state.reset()
 
     def test_notify_text_calls_provider(self):
-        from natlink_compat._ui_dispatch import notify_text
+        from natlink_compat._ui_protocol import notify_text
 
         mock_provider = MagicMock()
         self._state.ui_provider = mock_provider
@@ -139,7 +139,7 @@ class TestNotifyText(unittest.TestCase):
         mock_provider.on_text.assert_called_once_with("hello", 20)
 
     def test_notify_text_survives_provider_error(self):
-        from natlink_compat._ui_dispatch import notify_text
+        from natlink_compat._ui_protocol import notify_text
 
         bad_provider = MagicMock()
         bad_provider.on_text.side_effect = RuntimeError("boom")
@@ -149,7 +149,7 @@ class TestNotifyText(unittest.TestCase):
 
     def test_notify_text_reentrancy_guard(self):
         """Verify that re-entrant calls to notify_text are suppressed."""
-        from natlink_compat._ui_dispatch import notify_text, _guard
+        from natlink_compat._ui_protocol import notify_text, _guard
 
         calls = []
